% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{creates a dataset object for the decarb monitoring datasets}
\usage{
create_dataset(dataset_id)
}
\arguments{
\item{dataset_id}{id of the dataset}
}
\value{
list containing dataset-objects
}
\description{
A dataset object needs to be created for the download, the computation
as well as the publishing process
}
\details{
A dataset object inherits the following classes:
\itemize{
\item data_organization -> needed for the download process
\item dataset_id -> needed for the download as well as the publishing process
\item download_format -> needed for the download process
}
}
\concept{Datensatz erstellen}
