% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data.csv}
\alias{download_data.csv}
\title{Function to download data_format csv}
\usage{
\method{download_data}{csv}(ds)
}
\arguments{
\item{ds}{dataset object}
}
\description{
Function to download data_format csv
}
