% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Method dispatch for the data download}
\usage{
download_data(ds)
}
\arguments{
\item{ds}{dataset object to be processed}
}
\description{
Method dispatch for the data download
}
\concept{Download}
