% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data.xlsx}
\alias{download_data.xlsx}
\title{Function to download data_format xlsx}
\usage{
\method{download_data}{xlsx}(ds)
}
\arguments{
\item{ds}{dataset object}
}
\description{
Function to download data_format xlsx
}
