% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data.zip_csv}
\alias{download_data.zip_csv}
\title{Function to download zipped data files (csv)}
\usage{
\method{download_data}{zip_csv}(ds)
}
\arguments{
\item{ds}{dataset object}
}
\description{
Function to download zipped data files (csv)
}
