% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data_csv.openzh}
\alias{download_data_csv.openzh}
\title{Method specific to download csv form data_organization openzh
openzh refers to the Datenkatalog Kanton Zürich (https://www.zh.ch/daten)}
\usage{
\method{download_data_csv}{openzh}(ds)
}
\arguments{
\item{ds}{dataset object}
}
\description{
Method specific to download csv form data_organization openzh
openzh refers to the Datenkatalog Kanton Zürich (https://www.zh.ch/daten)
}
