% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data_xlsx.bfs}
\alias{download_data_xlsx.bfs}
\title{Method specific to download xlsx form data_organization BFS}
\usage{
\method{download_data_xlsx}{bfs}(ds)
}
\arguments{
\item{ds}{dataset object}
}
\description{
Method specific to download xlsx form data_organization BFS
}
