% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data_zip_csv.swisstopo}
\alias{download_data_zip_csv.swisstopo}
\title{Method specific to download zipped csv form data_organization swisstopo
The target file is the file to get from the zip folder}
\usage{
\method{download_data_zip_csv}{swisstopo}(ds)
}
\arguments{
\item{ds}{dataset object}
}
\description{
Method specific to download zipped csv form data_organization swisstopo
The target file is the file to get from the zip folder
}
