% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_helpers.R
\name{get_bfs_asset_info}
\alias{get_bfs_asset_info}
\title{Function to retrieve the asset number and the last year of the time series
for a specific BFS number}
\usage{
get_bfs_asset_info(ds)
}
\arguments{
\item{bfs_nr}{Number of a bfs publication e.g: "ind-d-21.02.30.1202.02.01"}
}
\description{
This function scrapes the asset page of a BFS-nr
!important: this bfs-nr here does not correspond to the bfs-nr of the municipality
}
