% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_helpers.R
\name{get_bfs_asset_nr}
\alias{get_bfs_asset_nr}
\title{Function to scrape the asset number of a specific bfs number}
\usage{
get_bfs_asset_nr(bfs_nr)
}
\arguments{
\item{bfs_nr}{Number of a bfs publication e.g: "ind-d-21.02.30.1202.02.01"}
}
\description{
This function scrapes the asset page of a bfs-nr for the asset number
!important: this bfs-nr is not the municipality bfs-nr
}
