% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_helpers.R
\name{get_download_url.bfs}
\alias{get_download_url.bfs}
\title{Method to create the download url for the XLSX data from the BFS
requires the asset number (BFS Nr) for the BFS DAM API}
\usage{
\method{get_download_url}{bfs}(ds)
}
\arguments{
\item{ds}{dataset object}
}
\description{
Method to create the download url for the XLSX data from the BFS
requires the asset number (BFS Nr) for the BFS DAM API
}
