% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_helpers.R
\name{get_download_url.csv}
\alias{get_download_url.csv}
\title{Method to create the download url for CSV data_type}
\usage{
\method{get_download_url}{csv}(ds)
}
\arguments{
\item{ds}{dataset object}
}
\description{
Method to create the download url for CSV data_type
}
