% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_helpers.R
\name{get_download_url.px}
\alias{get_download_url.px}
\title{Method to create the download url for PX and CSV data_type}
\usage{
\method{get_download_url}{px}(ds)
}
\arguments{
\item{ds}{dataset object}
}
\description{
Method to create the download url for PX and CSV data_type
}
