% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_helpers.R
\name{get_download_url.xlsx}
\alias{get_download_url.xlsx}
\title{Method to create the download url for the XLSX data from the BFS
requires the asset number (bfs nr) for the BFS DAM API}
\usage{
\method{get_download_url}{xlsx}(ds)
}
\arguments{
\item{ds}{dataset object}
}
\description{
Method to create the download url for the XLSX data from the BFS
requires the asset number (bfs nr) for the BFS DAM API
}
