% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_helpers.R
\name{get_download_url_xlsx.bfs}
\alias{get_download_url_xlsx.bfs}
\title{Method specific to create the download url for xlsx data_type from data_organization BFS}
\usage{
\method{get_download_url_xlsx}{bfs}(ds)
}
\arguments{
\item{ds}{dataset object}
}
\description{
Method specific to create the download url for xlsx data_type from data_organization BFS
}
