% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_helpers.R
\name{get_download_url_xlsx}
\alias{get_download_url_xlsx}
\title{Function that creates the download url for xlsx}
\usage{
get_download_url_xlsx(ds)
}
\description{
Function that creates the download url for xlsx
}
