% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_imports.R
\name{get_openzh_data}
\alias{get_openzh_data}
\title{Get data resource (csv) from open data ZH}
\usage{
get_openzh_data(file_name)
}
\arguments{
\item{file_name}{File name as provided at the resource on zh.ch/daten, e.g. KTZH_00001661_00003118.csv}
}
\value{
data.frame
}
\description{
Get data resource (csv) from open data ZH
}
