% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_imports.R
\name{get_pxdata}
\alias{get_pxdata}
\title{Get data from PXWEB}
\usage{
get_pxdata(px_cube, query_list)
}
\arguments{
\item{px_cube}{Path name of the data cube, e.g. "px-x-0103010000_102"}

\item{query_list}{Pre-constructed list element containing all query parameters}
}
\value{
Download data based on query list and convert to data.frame
}
\description{
Get data from PXWEB
}
