% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_helpers.R
\name{get_read_path.bfs}
\alias{get_read_path.bfs}
\title{Function for creating the FSP-specific read path}
\usage{
\method{get_read_path}{bfs}(ds)
}
\arguments{
\item{ds}{dataset object}
}
\description{
Use cases currently are XLSX via DAM API and via PXWEB data cubes
}
