% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_helpers.R
\name{get_read_path.default}
\alias{get_read_path.default}
\title{Default method used to create the download url for PX and CSV data_type
where the download url is a paste of url and id}
\usage{
\method{get_read_path}{default}(ds)
}
\arguments{
\item{ds}{dataset object}
}
\description{
Default method used to create the download url for PX and CSV data_type
where the download url is a paste of url and id
}
