% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_helpers.R
\name{get_read_path}
\alias{get_read_path}
\title{Function that creates the download url based on dataset_id and data_type}
\usage{
get_read_path(ds)
}
\arguments{
\item{ds}{dataset object}
}
\description{
Function that creates the download url based on dataset_id and data_type
}
