% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_helpers.R
\name{get_read_path_bfs.default}
\alias{get_read_path_bfs.default}
\title{Default method for creating the read path for data via the DAM API}
\usage{
\method{get_read_path_bfs}{default}(ds)
}
\description{
the asset number (BFS Nr) is required
}
