% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_helpers.R
\name{get_read_path_bfs.px}
\alias{get_read_path_bfs.px}
\title{Method for creating the read path for PXWEB data cubes}
\usage{
\method{get_read_path_bfs}{px}(ds)
}
\description{
the asset number (BFS Nr) is required
}
