% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_imports.R
\name{get_zip_data}
\alias{get_zip_data}
\title{Get zipped data files (csv) from opendata.swiss via api calls}
\usage{
get_zip_data(data_url, data_file_name)
}
\arguments{
\item{data_url}{Access url for the api}

\item{data_file_name}{File name including file format, e.g. RecycledWaste.csv}
}
\value{
data.frame
}
\description{
Get zipped data files (csv) from opendata.swiss via api calls
}
