% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_init.R
\name{indicator_init}
\alias{indicator_init}
\title{Initialisiere ein neues Skript für einen neuen Indikator}
\usage{
indicator_init(indicator_id)
}
\arguments{
\item{indicator_id}{ID of the indicator available in the indicator-list dataset}
}
\value{
script
}
\description{
Initialisiere ein neues Skript für einen neuen Indikator
}
\examples{

indicator_init("M3")

indicator_init("M100",ds_l)
}
