% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{read_data.default}
\alias{read_data.default}
\title{Default method}
\usage{
\method{read_data}{default}(ds)
}
\description{
The method is used in all cases where the input data is a csv, xlsx or
one of both packaged in a zipped folder
}
\details{
Working with import() from the rio package: https://cran.r-project.org/web/packages/rio/vignettes/rio.html
rio uses the file extension of a file name to determine what kind of file it is
and thus rio allows almost all common data formats to be read with the same function
!Attention: package seems to be stable. Nevertheless, keep an eye out for potential issues.
}
