% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{read_data.px}
\alias{read_data.px}
\title{Method to stream data from pxweb data cubes which is unique to the BFS
Downloads the data from a data cube based on a query list and converts it to a data.frame}
\usage{
\method{read_data}{px}(ds)
}
\description{
We are using the PXWEB package to stream BFS data:
https://ropengov.github.io/pxweb/articles/pxweb.html
}
\details{
!!! Important Limits: 10 calls per 10 sec., 5000  values per call

the name of the data cube, e.g. "px-x-0103010000_102", is taken from ds$data_id
the download url is constructed with the "get_read_path.bfs" method and added dto the ds as 'read_path'
the query list, a list element containing all query parameters, gets constructed with "get_px_query_list()"

To set up a PXWEB query list manually, start with a specific path and walk thorough each step
d <- pxweb::pxweb_interactive("https://www.pxweb.bfs.admin.ch/api/v1/de/px-x-0103010000_102")
}
