% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{read_data}
\alias{read_data}
\title{Function to stream data}
\usage{
read_data(ds)
}
\description{
For PXWEB data we need to define its own method, because we are working with
the PXWEB package for streaming data. Every other use case can be handled by
the default method.
}
