% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_data.R
\name{structure_data.Q1}
\alias{structure_data.Q1}
\title{method for Q1 - population data}
\usage{
\method{structure_data}{Q1}(ds)
}
\arguments{
\item{ds}{}
}
\description{
method for Q1 - population data
}
