% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_pipeline.R
\name{test_pipeline}
\alias{test_pipeline}
\title{Utility function to test if the pipeline works for all IDs}
\usage{
test_pipeline()
}
\value{
list object with all datasets and the data
}
\description{
Utility function to test if the pipeline works for all IDs
}
