% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Function for streaming the data.}
\usage{
download_data(ds)
}
\description{
We are only streaming the data, meaning we do not store local files of the
data. Data streamed is held temporarily and deleted afterwards.
}
\details{
The function first runs the get_read_path() method corresponding to the dataset ID,
with which the read path for the download URL is created.
Next the function uses the appropriate read_data() method to stream and append
the data to the dataset (ds) in the $data object.
}
