% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_helpers.R
\name{get_download_url}
\alias{get_download_url}
\title{Function that creates the download url based on dataset_id and data_type}
\usage{
get_download_url(ds)
}
\arguments{
\item{ds}{dataset object}
}
\description{
Function that creates the download url based on dataset_id and data_type
}
