% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_helpers.R
\name{get_read_path.default}
\alias{get_read_path.default}
\title{Standard method for creating the path of the download URL
for any data-holding organisation other than the FSO.}
\usage{
\method{get_read_path}{default}(ds)
}
\arguments{
\item{ds}{dataset object}
}
\description{
General path format: paste of base url and id
}
